#42: Move test files to test directory
