+++ test/test_setings.py
